hms_django/requirements.txt
Django==3.*
djangorestframework
pymysql
django-cors-headers
PyJWT
